refactor: reneme burger-item and burger-ingredient
