+++ src/components/burger-item/burger-item.module.css
@@ -1,15 +1,11 @@
-.card {
-	position: relative;
+.item {
+	width: 100%;
 	list-style-type: none;
-	width: 272px;
-	text-align: center;
-	display: flex;
-	flex-direction: column;
-
+	display: flex;	
 	align-items: center;
+	gap: 10px;
 }
 
-.image {
-	position: relative;
-	max-width: 240px;
+.item_without-icon {
+	justify-content: flex-end;
 }
