+++ src/components/ingredients/ingredients.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import cn from 'classnames';
-import BurgerItem from '../burger-item/burger-item';
+import BurgerIngredient from '../burger-ingredient/burger-ingredient';
 import styles from './ingredients.module.css';
 
 
@@ -10,7 +10,7 @@ const Ingredients = React.forwardRef(({ title, array, id, renderModal }, ref) =>
 		<section className={'mb-10'} >
 			<h2 id={id} ref={ref} className={cn(styles.ingredients__title, 'text', 'text_type_main-medium', 'mb-6')}>{title}</h2>
 			<ul className={cn(styles.list, 'ml-4')}>
-				{array.map((el) => <BurgerItem item={el} key={el._id} renderModal={renderModal} />)}
+				{array.map((el) => <BurgerIngredient item={el} key={el._id} renderModal={renderModal} />)}
 			</ul>
 		</section>
 	)
